+++ src/main/java/com/hyperfactions/migration/MigrationRegistry.java
@@ -3,6 +3,7 @@ package com.hyperfactions.migration;
 import com.hyperfactions.migration.migrations.config.ConfigV1ToV2Migration;
 import com.hyperfactions.migration.migrations.config.ConfigV2ToV3Migration;
 import com.hyperfactions.migration.migrations.config.ConfigV3ToV4Migration;
+import com.hyperfactions.migration.migrations.config.ConfigV4ToV5Migration;
 import org.jetbrains.annotations.NotNull;
 
 import java.nio.file.Path;
@@ -47,6 +48,7 @@ public class MigrationRegistry {
         register(new ConfigV1ToV2Migration());
         register(new ConfigV2ToV3Migration());
         register(new ConfigV3ToV4Migration());
+        register(new ConfigV4ToV5Migration());
     }
 
     /**